Finish itunes album tests & more enhancements

+++ internal/providers/itunes_provider.go
@@ -1,6 +1,6 @@
 package providers
 
 type ItunesResponseModel[T any] struct {
-	ResultCount int `json:"resultCount" validate:"required,gte=0"`
-	Results     []T `json:"results"     validate:"required,dive,required"`
+	ResultCount int64 `json:"resultCount" validate:"required,gte=0"`
+	Results     []T   `json:"results"     validate:"required,dive,required"`
 }
